Rename internet explorer package: iexplorer -> ie
