+++ src/main/java/com/github/nscuro/wdm/binary/ie/package-info.java
@@ -2,6 +2,6 @@
  * This package contains classes for binary management specific to Microsoft's Internet Explorer
  */
 @ParametersAreNonnullByDefault
-package com.github.nscuro.wdm.binary.iexplorer;
+package com.github.nscuro.wdm.binary.ie;
 
 import javax.annotation.ParametersAreNonnullByDefault;
